Passenger RecyclerView Select
This is where the driver can find all available passengers in the
vicinity, and choose to either /swipe LEFT/ or /checkmark and click
"Remove Passengers"/ to remove certain passengers. I will soon add a
"GO!" button or one similar, to press when the driver is ready to let
the app sort through the passengers for the best ones to take and
provide routes. The 'Route' tab/fragment is currently empty, that's
where the Google map will go.
NOTE: I haven't discovered how to uncheck all the checkboxes after
checking and deleting passengers. Recyclerview reuses containers of
deleted rows, so the new occupant of a deleted container continues to
have a checked checkbox. Will figure out how to fix this!